fix: load config file before parsing (#344)
supersedes #342
